增加函数 UniquePath 将路径转为唯一的路径（加当前时间戳）

+++ dofile/windows_test.go
@@ -12,6 +12,6 @@ func TestGetDriverInfo(t *testing.T) {
 		t.Fatal(err)
 	}
 
-	fmt.Println(dotext.BytesHumanReadable(free),
-		dotext.BytesHumanReadable(total), dotext.BytesHumanReadable(avail))
+	fmt.Println(dotext.BytesHumanReadable(int64(free)),
+		dotext.BytesHumanReadable(int64(total)), dotext.BytesHumanReadable(int64(avail)))
 }
